requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.3=pyhd8ed1ab_0
asttokens=2.2.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=openblas
bottleneck=1.3.5=py311hb9e55a9_0
bzip2=1.0.8=h1de35cc_0
ca-certificates=2023.01.10=hecd8cb5_0
certifi=2022.12.7=py311hecd8cb5_0
click=8.0.4=py311hecd8cb5_0
comm=0.1.3=pyhd8ed1ab_0
contourpy=1.0.7=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.6.6=py311h814d153_0
decorator=5.1.1=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
fonttools=4.39.3=pypi_0
greenlet=2.0.2=pypi_0
importlib-metadata=6.1.0=pyha770c72_0
importlib_metadata=6.1.0=hd8ed1ab_0
ipykernel=6.22.0=pyh736e0ef_0
ipython=8.11.0=pyhd1c38e8_0
ipywidgets=8.0.6=pypi_0
jedi=0.18.2=pyhd8ed1ab_0
joblib=1.2.0=pypi_0
jupyter_client=8.1.0=pyhd8ed1ab_0
jupyter_core=5.3.0=py311h6eed73b_0
jupyterlab-widgets=3.0.7=pypi_0
kiwisolver=1.4.4=pypi_0
libcxx=14.0.6=h9765a3e_0
libffi=3.4.2=hecd8cb5_6
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libopenblas=0.3.21=h54e7dc3_0
libsodium=1.0.18=hbcb3906_1
llvm-openmp=14.0.6=h0dcd299_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mysql-connector-python=8.0.32=pypi_0
mysqlclient=2.1.1=pypi_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
numexpr=2.8.4=py311hff9dab5_0
numpy=1.23.5=py311hff9dab5_0
numpy-base=1.23.5=py311h7183fb7_0
openssl=1.1.1t=hca72f7f_0
packaging=23.0=py311hecd8cb5_0
pandas=1.5.3=py311hc5848a5_0
parso=0.8.3=pyhd8ed1ab_0
patsy=0.5.3=pypi_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=pypi_0
pip=23.0.1=py311hecd8cb5_0
platformdirs=3.1.1=pyhd8ed1ab_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
protobuf=3.20.3=pypi_0
psutil=5.9.4=py311h5547dcb_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pygments=2.14.0=pyhd8ed1ab_0
pyparsing=3.0.9=pypi_0
python=3.11.0=h1fd4e5f_3
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=0.21.0=py311hecd8cb5_0
python_abi=3.11=2_cp311
pytz=2022.7=py311hecd8cb5_0
pyzmq=25.0.2=py311habfacb3_0
readline=8.2=hca72f7f_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
setuptools=65.6.3=py311hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlalchemy=2.0.9=pypi_0
sqlite=3.41.1=h6c40b1e_0
stack_data=0.6.2=pyhd8ed1ab_0
statsmodels=0.13.5=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h5d9f67b_0
tornado=6.2=py311h5547dcb_1
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
typing-extensions=4.5.0=hd8ed1ab_0
typing_extensions=4.5.0=pyha770c72_0
tzdata=2022g=h04d1e81_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.38.4=py311hecd8cb5_0
widgetsnbextension=4.0.7=pypi_0
xz=5.2.10=h6c40b1e_1
zeromq=4.3.4=he49afe7_1
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h4dc903c_0
